refactor(dashboard): reuse PAYG checkout
