+++ client/dashboard/src/components/billing/start-payg-checkout-cta.tsx
@@ -3,24 +3,15 @@ import {
   usePaygCheckoutAccess,
 } from "@/components/billing/payg-checkout-access";
 import {
-  isPaygCheckoutLocked,
-  setPaygCheckoutLocked,
-  usePaygCheckoutLocked,
-} from "@/components/billing/payg-checkout-lock";
+  PAYG_CHECKOUT_ERROR_MESSAGE,
+  useStartPaygCheckout,
+} from "@/components/billing/use-start-payg-checkout";
 import { Button } from "@/components/ui/Button";
 import { Text } from "@/components/ui/Text";
 import { ButtonSize } from "@/components/ui/lib/types";
-import { useTelemetry } from "@/contexts/Telemetry";
-import { safeExternalHttpUrl } from "@/lib/safe-external-url";
 import { cn } from "@/lib/utils";
-// Generated by the SDK for `usage.createStripeCheckout`, which lands
-// separately. If generation names the hook differently, only this import and
-// the `checkout.mutate` call below have to follow it.
-import { useCreateStripeCheckoutMutation } from "@gram/client/react-query/createStripeCheckout.js";
-import { useCallback, useState } from "react";
 
 const CTA_LABEL = "Start pay as you go";
-const ERROR_MESSAGE = "Couldn't start checkout. Try again.";
 
 /**
  * Puts an organization onto pay as you go by handing the admin a Stripe
@@ -49,63 +40,8 @@ export function StartPaygCheckoutCTA({
   eligibility?: PaygCheckoutEligibility;
 }): JSX.Element | null {
   const { eligible, activeOrganizationId } = usePaygCheckoutAccess(eligibility);
-  const telemetry = useTelemetry();
-  // The lock is released by the mutation's own callbacks rather than ones
-  // passed to `mutate`: React Query skips the per-call callbacks when the mount
-  // that dispatched them unmounts while the request is still in flight, which
-  // would leave the organization locked for every later mount.
-  //
-  // The organization travels through the mutation context instead of a closure
-  // because React Query rewrites a pending mutation's options on every render:
-  // switching organizations mid-checkout would otherwise release the lock on
-  // the organization switched *to* and wedge the one that started it. The
-  // context is captured once, when the mutation starts.
-  const checkout = useCreateStripeCheckoutMutation({
-    onMutate: () => activeOrganizationId,
-    onSettled: (_data, _error, _variables, context) => {
-      if (typeof context === "string") {
-        setPaygCheckoutLocked(context, false);
-      }
-    },
-  });
-  const [failed, setFailed] = useState(false);
-  // `isPending` only disables the button on the NEXT render, and only for the
-  // mount that owns the mutation, so a second click — on either mount —
-  // dispatched before React re-renders would open a second checkout session.
-  // The lock closes that window synchronously for every mount.
-  const locked = usePaygCheckoutLocked(activeOrganizationId);
-
-  const startCheckout = useCallback(() => {
-    if (isPaygCheckoutLocked(activeOrganizationId)) return;
-    setPaygCheckoutLocked(activeOrganizationId, true);
-    setFailed(false);
-
-    checkout.mutate(
-      {},
-      {
-        onSuccess: (link) => {
-          const url = safeExternalHttpUrl(link);
-          if (url === null) {
-            telemetry.capture("payg_checkout_error", {
-              error: "unusable checkout link",
-            });
-            setFailed(true);
-            return;
-          }
-
-          // Stripe returns the customer to the dashboard when checkout
-          // finishes, so this is a handoff rather than a new tab.
-          window.location.assign(url);
-        },
-        onError: (error: unknown) => {
-          telemetry.capture("payg_checkout_error", {
-            error: error instanceof Error ? error.message : "unknown",
-          });
-          setFailed(true);
-        },
-      },
-    );
-  }, [activeOrganizationId, checkout, telemetry]);
+  const { startCheckout, isPending, error } =
+    useStartPaygCheckout(activeOrganizationId);
 
   if (!eligible) return null;
 
@@ -115,13 +51,13 @@ export function StartPaygCheckoutCTA({
         variant="primary"
         size={size}
         onClick={startCheckout}
-        disabled={locked || checkout.isPending}
+        disabled={isPending}
       >
         {label}
       </Button>
-      {failed ? (
+      {error ? (
         <Text small destructive role="alert">
-          {ERROR_MESSAGE}
+          {PAYG_CHECKOUT_ERROR_MESSAGE}
         </Text>
       ) : null}
     </div>
